requirements.txt
argcomplete==1.8.2
base58==0.2.4
boto3==1.4.5
botocore==1.5.40
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
docutils==0.14
durationpy==0.5
Flask==0.12.2
Flask-Cors==3.0.3
future==0.16.0
futures==3.1.1
hjson==3.0.0
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.16.1
MarkupSafe==1.0
placebo==0.8.1
python-dateutil==2.6.1
python-dotenv==0.6.5
python-slugify==1.2.4
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.10
six==1.10.0
toml==0.9.2
tqdm==4.15.0
troposphere==1.9.5
Unidecode==0.4.21
urllib3==1.22
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.43.2
